Update patch set 2
Patch Set 2:

here is the problem, to build the documentation, sphinx tries to load openstack_dashboard.local.local_settings. It can't because it's a symlink to /etc/openstack_dashboards/local_settings and this file can only be read by root:apache (640). Even if the bytecode .pyc has been created, in python3 it always tries to stat the .py file, so it can't load the LOCAL_PATH configuration which is in /etc/openstack_dashboard/local_settings.

The best solution i can figure out is give read permissions to other in /etc/openstack_dashboard/local_settings in horizon package but i need to check if there is any security sensitive info there.

Patch-set: 2