Update comment to avoid = character
When rdopkg adds a patch to a spec file, it tries to find what it
calls a "magic comment", being two lines starting with a #, and
the second one containing "something=something" in it.

Let's update the comments in the spec to avoid this corner case, so
rdopkg does not get confused.

Change-Id: I742101edf6e215c81302ee3c5d298e0525a53afd